Ajout de classes communes
