+++ sachem/Méthodes Communes/Comparateur.cs
