+++ sachem/Méthodes Communes/Crypt.cs
