+++ sachem/Méthodes Communes/ValidationsAcces.cs
