+++ sachem/sachem.csproj
@@ -218,14 +218,13 @@
     <Compile Include="App_Start\FilterConfig.cs" />
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="App_Start\Startup.Auth.cs" />
-    <Compile Include="Classes_Sachem\Comparateur.cs" />
-    <Compile Include="Classes_Sachem\Liste.cs" />
-    <Compile Include="Classes_Sachem\ValidationsAcces.cs" />
+    <Compile Include="Méthodes Communes\Comparateur.cs" />
+    <Compile Include="Méthodes Communes\ValidationsAcces.cs" />
     <Compile Include="Controllers\AccountController.cs" />
     <Compile Include="Controllers\ConsulterCoursController.cs" />
     <Compile Include="Controllers\ContratEngagementController.cs" />
     <Compile Include="Controllers\CoursController.cs" />
-    <Compile Include="Classes_Sachem\Crypt.cs" />
+    <Compile Include="Méthodes Communes\Crypt.cs" />
     <Compile Include="Controllers\CoursSuiviController.cs" />
     <Compile Include="Controllers\DossierEtudiantController.cs" />
     <Compile Include="Controllers\EnseignantController.cs" />
@@ -372,7 +371,7 @@
     <Compile Include="Models\ReponseQuestion.cs">
       <DependentUpon>Model1.tt</DependentUpon>
     </Compile>
-    <Compile Include="Models\SACHEMIdentite.cs" />
+    <Compile Include="Méthodes Communes\SACHEMIdentite.cs" />
     <Compile Include="Models\sachemModelDN.cs" />
     <Compile Include="Models\SACHEMModels.cs" />
     <Compile Include="Models\Section.cs">
